+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.8] - 2025-06-29
+
+### Changed
+
+- Unique fraction at 1000 benchmark metric
+- Scores now match Moses benchmark scores except for scaffold similarity
+
 ## [0.1.6] - 2025-06-28
 
 ### Changed